+++ 01-domain/src/main/scala/camundala/domain/MockedServiceResponse.scala
@@ -67,9 +67,9 @@ object MockedServiceResponse:
     MockedServiceResponse(status, Left(None))
 
   given tapirSchema[ServiceOut: Schema]: Schema[MockedServiceResponse[ServiceOut]] =
-    Schema.derived[MockedServiceResponse[ServiceOut]]
+    deriveSchema[MockedServiceResponse[ServiceOut]]
 
-  given mockedHttpResponseEncoder[ServiceOut: Encoder]: Encoder[MockedServiceResponse[ServiceOut]] =
+  given mockedHttpResponseJsonEncoder[ServiceOut: JsonEncoder]: JsonEncoder[MockedServiceResponse[ServiceOut]] =
     Encoder.instance { response =>
       Json.obj(
         "respStatus" -> Json.fromInt(response.respStatus),
@@ -84,7 +84,7 @@ object MockedServiceResponse:
       )
     }
 
-  given mockedHttpResponseDecoder[ServiceOut: Decoder]: Decoder[MockedServiceResponse[ServiceOut]] =
+  given mockedHttpResponseJsonDecoder[ServiceOut: JsonDecoder]: JsonDecoder[MockedServiceResponse[ServiceOut]] =
     Decoder.instance { cursor =>
       for
         respStatus <- cursor.downField("respStatus").as[Int]