+++ 03-dmn/src/main/scala/camundala/dmn/DmnTesterConfigCreator.scala
@@ -1,6 +1,7 @@
 package camundala
 package dmn
 
+import camundala.domain.*
 import camundala.bpmn.*
 import pme123.camunda.dmn.tester.shared.*