+++ 05-examples/demos/src/main/scala/camundala/examples/demos/SimulationTestOverridesExample.scala
@@ -8,8 +8,8 @@ object SimulationTestOverridesExample extends BpmnDsl:
   case class SimpleObject(name: String = "salu", other: Boolean = false)
 
   object SimpleObject:
-    given Schema[SimpleObject] = Schema.derived
-    given CirceCodec[SimpleObject] = deriveCodec
+    given ApiSchema[SimpleObject] = deriveSchema
+    given JsonCodec[SimpleObject] = deriveCodec
 
   case class InOutput(
                        simpleValue: String = "hello",
@@ -18,8 +18,8 @@ object SimulationTestOverridesExample extends BpmnDsl:
                        objectCollectionValue: Seq[SimpleObject] = Seq(SimpleObject(), SimpleObject("tschau", true)),
                      )
   object InOutput:
-    given Schema[InOutput] = Schema.derived
-    given CirceCodec[InOutput] = deriveCodec
+    given ApiSchema[InOutput] = deriveSchema
+    given JsonCodec[InOutput] = deriveCodec
 
   lazy val simulationProcess = process(
     id = "simulation-TestOverrides",