+++ 05-examples/demos/src/main/scala/camundala/examples/demos/VariablesExample.scala
@@ -17,8 +17,8 @@ object VariablesExample extends BpmnDsl:
   )
 
 
-  given Schema[Input] = Schema.derived
-  given CirceCodec[Input] = deriveCodec
+  given ApiSchema[Input] = deriveSchema
+  given JsonCodec[Input] = deriveCodec
 
 
 end VariablesExample