+++ 05-examples/invoice/camunda7/src/main/scala/camundala/examples/invoice/exports.scala
@@ -2,9 +2,10 @@ package camundala.examples.invoice
 
 import camundala.domain.*
 
-enum InvoiceCategory derives ConfiguredEnumCodec:
+enum InvoiceCategory:
   case `Travel Expenses`, Misc, `Software License Costs`
 
 object InvoiceCategory:
-  given Schema[InvoiceCategory] = Schema.derived
+  given JsonCodec[InvoiceCategory] = deriveCodec
+  given ApiSchema[InvoiceCategory] = deriveSchema
 end InvoiceCategory